initial implementation
Does not reflect the subcommand structure, but otherwise is fairly
complete.

Settings that can be configured for scans are defined in a JSON file
(`wordfence/configconfig_definitions.json`)

There are two main deficiencies:

Files passed in as positional arguments after settings are not currently
added to the Config object.

The second shortcoming is that flags, once inverted by the INI file,
cannot be inverted again by the CLI options of the config. For
example, if `output=yes` is defined in the INI file, the CLI command has
no way of turning the flag off.

However, if the INI file sets the flag to the default value (`output=no`
in this example), the CLI command can still invert the default.

This will need attention eventually but may not be a blocker for the
MVP.